Exception overhaul, API upgrade
Not being able to return a default image isn't catastrophic and is
probably best handled by the consumer. As such, the FileNotFound
exception is caught and but logged.

Other similar changes have been made elsewhere in the code, as well as
providing enhanced error messages for when things go wrong (e.g.,
trying to upload a missing file).

API now allows for the deletion of files and the swapping out of
default file based on file name.